adjust routes and component names to break out dashboard list and dashboard
